fixed Wire.read() return type and value when no data in buffer

+++ pic32/libraries/Wire/Wire.h
@@ -107,7 +107,7 @@ class TwoWire
     int write(char*);
     uint8_t available(void);
     uint8_t __attribute__((deprecated("Use read() instead"))) receive(void);
-    uint8_t read(void);
+    int read(void);
     void onReceive( void (*)(int) );
     void onRequest( void (*)(void) );
 };
